chore: corrections to PR#5
Integer id -> Long id

+++ spring-05/src/main/java/ru/korolkovrs/spring/dao/AuthorDao.java
@@ -5,7 +5,7 @@ import ru.korolkovrs.spring.domain.Author;
 import java.util.List;
 
 public interface AuthorDao {
-    Author getById(Integer id);
+    Author getById(Long id);
 
     Author getByName(String name);
